Adapt project structure to contain Kiali application
