--- plugin/src/components/MeshTab/index.ts
@@ -1,4 +1,0 @@
-export * from './IstioMesh';
-export * from './ProjectMesh';
-export * from './ServiceMesh';
-export * from './WorkloadMesh';
