+++ plugin/src/kiali/README.md
@@ -1,0 +1,1 @@
+Kiali application code will be copied here.
