+++ plugin/src/openshift/pages/MeshTab/IstioMesh.tsx
@@ -1,11 +1,13 @@
 import * as React from 'react';
-import userProps, {getKialiUrl, initKialiListeners, kioskUrl} from '../../kialiIntegration';
+import userProps, {getKialiUrl, initKialiListeners, kioskUrl} from '../../utils/KialiIntegration';
 
-type ProjectMeshProps = {
+type IstioMeshProps = {
+    configType: string;
+    namespace: string;
     idObject: string;
 }
 
-export const ProjectMesh = (props: ProjectMeshProps) => {
+export const IstioMesh = (props: IstioMeshProps) => {
     const [kialiUrl, setKialiUrl] = React.useState({
         baseUrl: '',
         token: '',
@@ -19,8 +21,8 @@ export const ProjectMesh = (props: ProjectMeshProps) => {
             .catch(e => console.error(e));
     }, []);
 
-    const iFrameUrl = kialiUrl.baseUrl +  '/console/graph/namespaces?namespaces=' + props.idObject + '&' + kioskUrl() + '&' + kialiUrl.token
-    + '&duration=' + userProps.duration + '&timeRange=' + userProps.timeRange;
+    const iFrameUrl = kialiUrl.baseUrl + '/console/namespaces/' + props.namespace + '/istio/' + props.configType + '/' + props.idObject + '?' + kioskUrl() + '&'
+    + kialiUrl.token + '&duration=' + userProps.duration + '&timeRange=' + userProps.timeRange;
     return (
         <iframe
                 src={iFrameUrl}
