+++ plugin/src/openshift/utils/IstioResources.ts
